Sections are now moving lens
This is the setup for eventually moving the content inside the lens. For now, showing that I can move it until I have images for the section backgrounds and inside the lens.

I will have to address the restart of the animation rather than resuming while passing into a new section. IN the future this will move the background image up or down depending on travel.